app/requirements.txt
flask
flask-pymongo
gunicorn
python-dotenv